+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Automatically suppresses UI Fixes' magazine retention reload feature
 * Added a notification when you lose a magazine during reload
 * Added a setting to disable this notification
-* Added a setting to only retain magazines if the weapon's mastering is maxed out
+* Added a setting to only retain magazines if the weapon's mastering is maxed-out
 * Added a setting to disable this mod's functionality entirely
 * Fixed magazines dropping when they no longer fit the slot the new magazine came from